functions/Images/requirements.txt
functions-framework==3.*
fastapi==0.109.0
uvicorn==0.24.0
google-cloud-storage==2.14.0
